+++ example/lib/main.dart
@@ -1,17 +1,19 @@
+import 'package:flutter/widgets.dart';
 import 'package:flutter/material.dart';
 import 'dart:async';
-
 import 'package:xfvoice/xfvoice.dart';
 
 void main() => runApp(MyApp());
 
 class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
+  _MyApp createState() => _MyApp();
 }
 
-class _MyAppState extends State<MyApp> {
-  String iflyResultString = '按下开始识别，松手结束识别';
+class _MyApp extends State<MyApp> {
+  String voiceMsg = '暂无数据';
+  String iflyResultString = '按下方块说话';
+
+  XFJsonResult xfResult;
 
   @override
   void initState() {
@@ -19,59 +21,81 @@ class _MyAppState extends State<MyApp> {
     initPlatformState();
   }
 
-  // Platform messages are asynchronous, so we initialize in an async method.
-  Future<void> initPlatformState() async {
+   Future<void> initPlatformState() async {
     final voice = XFVoice.shared;
-    voice.init(appIdIos: '5d133a41', appIdAndroid: '5d199f2d');
+    // 请替换成你的appid
+    voice.init(appIdIos: '5d133a41', appIdAndroid: '5d133aae');
     final param = new XFVoiceParam();
     param.domain = 'iat';
-    param.asr_ptt = '0';
-    param.asr_audio_path = 'xme.pcm';
-    param.result_type = 'plain';
-    voice.setParameter(param.toMap());
+    // param.asr_ptt = '0';   //取消注释可去掉标点符号
+    param.asr_audio_path = 'audio.pcm';
+    param.result_type = 'json'; //可以设置plain
+    final map = param.toMap();
+    map['dwa'] = 'wpgs';        //设置动态修正，开启动态修正要使用json类型的返回格式
+    voice.setParameter(map);
   }
 
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      title: '测试的demo',
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: new Text('测试demo'),
         ),
         body: Center(
           child: GestureDetector(
-            child: Text(iflyResultString),
-            onTapDown: onTapDown,
-            onTapUp: onTapUp,
+            child: Container(
+              child: Text(iflyResultString),
+              width: 300.0,
+              height: 300.0,
+              color: Colors.blueAccent,
+            ),
+            onTapDown: (d) {
+              setState(() {
+                voiceMsg = '按下';
+              });
+              _recongize();
+            },
+            onTapUp: (d) {
+              // _recongizeOver();
+            },
           ),
-        ),
+        )
       ),
     );
   }
 
-  onTapDown(TapDownDetails detail) {
-    iflyResultString = '';
+  void _recongize() {
     final listen = XFVoiceListener(
       onVolumeChanged: (volume) {
-        print('$volume');
+      },
+      onBeginOfSpeech: () {
+        xfResult = null;
       },
       onResults: (String result, isLast) {
+        if (xfResult == null) {
+          xfResult = XFJsonResult(result);
+        } else {
+          final another = XFJsonResult(result);
+          xfResult.mix(another);
+        }
         if (result.length > 0) {
           setState(() {
-            iflyResultString += result;
+            iflyResultString = xfResult.resultText();
           });
         }
       },
       onCompleted: (Map<dynamic, dynamic> errInfo, String filePath) {
         setState(() {
-          iflyResultString += '\n$filePath';
+          
         });
       }
     );
     XFVoice.shared.start(listener: listen);
   }
 
-  onTapUp(TapUpDetails detail) {
+  void _recongizeOver() {
     XFVoice.shared.stop();
   }
 }