+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 ## 1.0.1
 
 * fix: result is null on last in iOS.
+
+## 1.0.2
+
+* fix: combine result Strings in Android.
+
+## 1.0.3
+
+* fix: adapt the json result and add accent support.